add block ip

+++ controller/participantController.js
@@ -27,7 +27,11 @@ var ip = [
 
 // Handle index actions
 exports.index = function (req, res) {
-  
+  if (!ip.includes(req.ip.replace("::ffff:", ""))) {
+    console.log(req.ip.replace("::ffff:", ""));
+
+    return res.status(500).send();
+  }
   Participant.get(function (err, participants) {
     if (err) {
       return res.json({
@@ -111,7 +115,11 @@ exports.view = function (req, res) {
 
 // Handle create actions
 exports.new = function (req, res) {
-  
+  if (!ip.includes(req.ip.replace("::ffff:", ""))) {
+    console.log(req.ip.replace("::ffff:", ""));
+
+    return res.status(500).send();
+  }
 
   var participant = new Participant();
   participant.name = req.body.name;
@@ -147,7 +155,11 @@ exports.new = function (req, res) {
 
 // Handle update actions
 exports.update = function (req, res) {
-  
+  if (!ip.includes(req.ip.replace("::ffff:", ""))) {
+    console.log(req.ip.replace("::ffff:", ""));
+
+    return res.status(500).send();
+  }
 
   var moveSession = false;
   var oldSession = {};
@@ -265,7 +277,12 @@ exports.vote = function (req, res) {
 
 // Handle delete actions
 exports.delete = function (req, res) {
-  
+  if (!ip.includes(req.ip.replace("::ffff:", ""))) {
+    console.log(req.ip.replace("::ffff:", ""));
+
+    return res.status(500).send();
+  }
+
   Participant.findById(req.params.id, function (err, participant) {
     if (err) return res.send(err);
 
@@ -299,7 +316,11 @@ exports.delete = function (req, res) {
 
 // Handle delete actions
 exports.force_delete = function (req, res) {
-  
+  if (!ip.includes(req.ip.replace("::ffff:", ""))) {
+    console.log(req.ip.replace("::ffff:", ""));
+
+    return res.status(500).send();
+  }
   Participant.deleteOne(
     {
       _id: req.params.id,
